requirements.txt
streamlit
requests
together
python-dotenv
Pillow
googlesearch-python